% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_peaks_utils.R
\name{define_domain_per_peak}
\alias{define_domain_per_peak}
\title{Define the peak domains}
\usage{
define_domain_per_peak(
  np,
  C2,
  C2_peaks_pos,
  C2_peaks_thresholded_pos,
  spanD.left,
  spanD.right,
  spread = 3
)
}
\arguments{
\item{np}{peak number}

\item{C2}{channel intensity}

\item{C2_peaks_pos}{positions of peaks}

\item{C2_peaks_thresholded_pos}{positions of peaks above threshold}

\item{spanD.left}{left span of domain}

\item{spanD.right}{right span of domain}

\item{spread}{parameter}
}
\value{
a data.frame with the domains around peak
}
\description{
Define the peak domains
}
