% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_peaks_utils.R
\name{find_peaks}
\alias{find_peaks}
\title{Find the peaks positions}
\usage{
find_peaks(C2)
}
\arguments{
\item{C2}{channel intensity}
}
\value{
peak positions
}
\description{
Find the peaks positions
}
