% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_peaks_utils.R
\name{get_threshold}
\alias{get_threshold}
\title{Determine threshold}
\usage{
get_threshold(C2)
}
\arguments{
\item{C2}{channel intensity}
}
\value{
a threshold for `C2`
}
\description{
Determine threshold
}
