% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_peaks_utils.R
\name{intensity_under_curve}
\alias{intensity_under_curve}
\title{Compute area under curve}
\usage{
intensity_under_curve(x, y, method = c("trapz", "sum"))
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{method}{integration method}
}
\value{
The area under the curve (single numeric value)
}
\description{
Compute area under curve
}
\details{
If method is `trapz`, uses the Trapezoidal Integration from `{pracma}`. If method is `sum`, simply
sums the heights of `y` (ignoring `x` altogether).
}
