% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_peaks_utils.R
\name{interpeak_dist}
\alias{interpeak_dist}
\title{Compute distance between peaks}
\usage{
interpeak_dist(x_peaks, n)
}
\arguments{
\item{x_peaks}{positions of peaks}

\item{n}{number of pixels}
}
\value{
a vector of length `length(x_peaks)-1` with the distances between peaks
}
\description{
Compute distance between peaks
}
