% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_one_thresholded}
\alias{plot_one_thresholded}
\title{Plot 1 individual's profile with threshold}
\usage{
plot_one_thresholded(C2, C2_peaks_thresholded_pos, threshold)
}
\arguments{
\item{C2}{intensity channel}

\item{C2_peaks_thresholded_pos}{positions of peaks}

\item{threshold}{threshold}
}
\description{
Plot 1 individual's profile with threshold
}
