% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_one_worm}
\alias{plot_one_worm}
\title{Plot 1 individual's profile
Title}
\usage{
plot_one_worm(data, ind_name = "#1")
}
\arguments{
\item{data}{data.frame for a all individuals, with columns `individual`, `values.C2`, `values.C1`}

\item{ind_name}{one value in `data$individual` to filter on}
}
\value{
invisibly, the input data
}
\description{
Plot 1 individual's profile
Title
}
