% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_whole_directory.R
\name{process_whole_directory}
\alias{process_whole_directory}
\title{Run whole script}
\usage{
process_whole_directory(
  input_dir,
  output_dir,
  plot_path = NULL,
  dim.png = c(1000, 600)
)
}
\arguments{
\item{input_dir}{where to find the csv files}

\item{output_dir}{where to save results}

\item{plot_path}{either NULL (do not save plots) or a path to a directory to save plots}

\item{dim.png}{dimensions (in pixels) of the plots to save, ignored if plot_path = FALSE}
}
\value{
Saves in the output_dir, invisibly returns the same data
}
\description{
Run whole script
}
\details{
Given a directory of csv files, exports coefficients.

This is equivalent to reading the files with `read_condition()`, then, for each individual,
run `threshold_peaks_one_individual()`, and write the compiled data. The individual functions can
also be called directly, see their help pages.
}
