% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_peaks.R
\name{threshold_peaks_one_indiv}
\alias{threshold_peaks_one_indiv}
\title{Threshold the peaks for single individual}
\usage{
threshold_peaks_one_indiv(data_filt, plot_path = FALSE, dim.png = c(1000, 600))
}
\arguments{
\item{data_filt}{data.frame subsetted for one individual}

\item{plot_path}{either FALSE (do not save plots) or a path to a directory to save plots}

\item{dim.png}{dimensions of the plot to save, ignored if plot_path = FALSE}
}
\value{
a data.frame with 1 row, and columns for all computed variables
}
\description{
Threshold the peaks for single individual
}
