% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_peaks.R
\name{threshold_peaks_one_individual}
\alias{threshold_peaks_one_individual}
\title{Threshold the peaks for single individual}
\usage{
threshold_peaks_one_individual(
  data_filt,
  column_names = "individual",
  plot_path = NULL,
  dim.png = c(1000, 600),
  method_integration = "trapz"
)
}
\arguments{
\item{data_filt}{data.frame subsetted for one individual}

\item{column_names}{name of the column containing the individual names}

\item{plot_path}{either NULL (do not save plots) or a path to a directory to save plots}

\item{dim.png}{dimensions of the plot to save, ignored if plot_path = FALSE}

\item{method_integration}{method to determine area under the intensity curve within the domains}
}
\value{
a data.frame with 1 row, and columns for all computed variables
}
\description{
Threshold the peaks for single individual
}
